Extracted json file download logic..
..so that it's easier to mock in tests.
